+++ CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-* Added: Command to add a template to an adb item
-* Added: Command to add a parent to an adb item
+* Added: Command to add and remove templates from the adb item
+* Added: Command to add and remove parents from the adb item
 
 ## 1.3.0 - 2020-04-08